LICENSE file required
Sir,
Without a license file, a user is legally not allowed to copy your code and run it on any device, We would like to get this code Copyrighted but allow other users to have a copy of it. This [license](https://en.wikipedia.org/wiki/MIT_License) is my favorite. It commands the user that when downloading the code, a copy of it must be in the files! This LICENSE file affirms you are the creator of this work, and this is your right that this belongs to you and is your original work.